% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeFunctions.R
\name{cbind.zoon}
\alias{cbind.zoon}
\title{cbind.zoon}
\usage{
cbind.zoon(a, b)
}
\arguments{
\item{a}{The dataframe of which you'd like preserve the user defined attributes}

\item{b}{The dataframe which you'd like to append to a.}
}
\description{
cbind.zoon
}

