% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeFunctions.R
\name{cbindZoon}
\alias{cbindZoon}
\title{cbindZoon}
\usage{
cbindZoon(a, b)
}
\arguments{
\item{a}{The dataframe of which you'd like preserve the user defined
attributes}

\item{b}{The dataframe which you'd like to append to a.}
}
\description{
cbindZoon
}
