% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChangeWorkflow.R
\name{ChangeWorkflow}
\alias{ChangeWorkflow}
\title{Change a workflow and rerun.}
\usage{
ChangeWorkflow(workflow, occurrence = NULL, covariate = NULL,
  process = NULL, model = NULL, output = NULL, forceReproducible = NULL)
}
\arguments{
\item{workflow}{A zoonWorkflow object from a previous zoon analysis}
}
\value{
A list with the results of each module and a copy of the
 call used to execute the workflow (
}
\description{
Takes a workflow object and reruns it with changes.
}
\examples{
\dontrun{
w <- workflow(UKAnophelesPlumbeus, UKAir,
              OneHundredBackground,
              LogisticRegression,
              SameTimePlaceMap)

ChangeWorkflow(w)
}
}

