% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CheckModList}
\alias{CheckModList}
\title{CheckModList}
\usage{
CheckModList(x)
}
\description{
Helper to sort module arguments into lists with common structure.
  Want any input to end up as
  list(module = moduleName, paras = list(paraName1 = para, paraName2 = 2)).
  See tests/testthat/testZoon.R for list of potential input.
}

