% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{CWBZim}
\alias{CWBZim}
\title{Presence/absence of the coffee white stem borer in Zimbabwe 2003}
\format{data.frame with five columns, longitude, latitude, value (1 for
   presence), type (presence) and a column of 1s indicating this is
   training data not external validation data.}
\source{
Original publication:
Kutywayo D, Chemura A, Kusena W, Chidoko P, Mahoya C (2013) The impact of
 climate change on the potential distribution of agricultural pests: the case
  of the coffee white stem borer (\emph{Monochamus leuconotus} P.) in
   Zimbabwe. PLoS ONE 8(8): e73432.
   Dryad data package:
   Kutywayo D, Chemura A, Kusena W, Chidoko P, Mahoya C (2013) Data from:
   The impact of climate change on the potential distribution of agricultural
    pests: the case of the coffee white stem borer (\emph{Monochamus
     leuconotus} P.) in Zimbabwe. Dryad Digital Repository.
   \url{http://dx.doi.org/10.1371/journal.pone.0073432}
}
\description{
This is an example presence/absence dataset for the
coffee white stem borer \emph{Monochamus leuconotus} P. taken from an
open access dataset on the Dryad data repository.
The data are made available by those authors under a Creative Commons CC0
These data are used in the module CWBZimbabwe which can be used for running
toy presence/absence species distribution models.
}
\author{
Nick Golding August 2015
}
