% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ErrorAndSave}
\alias{ErrorAndSave}
\title{ErrorAndSave
Function used in tryCatch calls in workflow.
  If an error is caught, save the workflow to tmpZoonWorkflow.
  And return some useful messages. Then stop().
  cond is the error messages passed by try catch.
  mod is the modules number (1:5) to give NULLS to the correct modules.}
\usage{
ErrorAndSave(cond, mod = 1, e)
}
\arguments{
\item{cond}{The error message that was caught in tryCatch.}

\item{mod}{Which module has failed? 1=occurrence, 2=covariate, 3=process
4=model, 5=output.}

\item{e}{The workflow call environment}
}
\description{
ErrorAndSave
Function used in tryCatch calls in workflow.
  If an error is caught, save the workflow to tmpZoonWorkflow.
  And return some useful messages. Then stop().
  cond is the error messages passed by try catch.
  mod is the modules number (1:5) to give NULLS to the correct modules.
}

