% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractAndCombData}
\alias{ExtractAndCombData}
\title{ExtractAndCombData}
\usage{
ExtractAndCombData(occurrence, ras)
}
\arguments{
\item{occurrence}{A data frame from an occurrence module}

\item{ras}{Environmental raster layer, stack or brick.}
}
\description{
Simply extract covariates from rasters and combine with occurrence data.
}

