% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindExtent.R
\name{FindExtent}
\alias{FindExtent}
\title{Determine the Extent of Interest}
\usage{
FindExtent(
  initial_extent = c(-180, 180, -90, 90),
  resolution = c("low", "medium"),
  round = 3
)
}
\arguments{
\item{initial_extent}{optional numeric vector or extent object defining the
giving the extent (in latitude and longitude) of the world map to plot.
This should be larger than the target extent so that you can click within
it. Can be useful for 'zooming in' to an area to define an extent more
precisely.}

\item{resolution}{how detailed to make national borders in the plotted world
map. "low" is less accurate, but faster to load than "medium"}

\item{round}{to how many decimal places the extent should be reported. The
default value rounds to 3 decimal places. Set this to \code{Inf} to prevent
any rounding. This only affects the vector version of the extent printed in
the console, not the extent object returned.}
}
\value{
invisibly returns an extent object, also prints a vector version of
  that extent to the console.
}
\description{
Helper function to define an extent (in latitude and longitude)
  describing the area of interest for modelling. Opens a world map, on which
  you can click to twice to determine the area of interest.
}
\details{
This is just a thin wrapper around \code{raster::getExtent},
  providing the world map to click on and reporting the extent as a rounded
  vector.
}
\examples{
\dontrun{
 # open a world map and click on two spots to print the extent to the console
 FindExtent()

 # you can get the corresponding extent object too
 ext <- FindExtent()
 ext

 # you can zoom in on an area and increase the resolution, and precision of
 # the extent vector
 FindExtent(c(112, 156, -44, -8),
            resolution = "medium",
            round = 6)
}
}
