% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FrescaloBias}
\alias{FrescaloBias}
\title{Example bias raster for plants in England}
\format{RasterLayer with extent: c(-7.083, 2.167, 49.83, 55.83) and values ranging between 0 and 1.}
\description{
This is an example bias raster giving a modelled estimate of the
relative recording effort for plants in England using the Frescalo function in the R package sparta.
}
\author{
Tom August & Nick Golding September 2015
}
