% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{GetMaxEnt}
\alias{GetMaxEnt}
\title{Get MaxEnt}
\usage{
GetMaxEnt()
}
\description{
Helper function to get the MaxEnt java executable and install it in
 the right locations for zoon modules that use `dismo::maxent` and
 `biomod2`.
}
\details{
Since MaxEnt may not be distributed other than via the MaxEnt
website, users must download the file themselves and place it in the
right location for R packages to access it. This function helps with that.
Just run \code{GetMaxEnt()} and follow the prompts in the terminal.
}
