% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetModuleList}
\alias{GetModuleList}
\title{Get a list of all the modules available on the github repo.}
\usage{
GetModuleList()
}
\value{
A list with all module names.
}
\description{
Get a list of all the modules available on the github repo.
}
\examples{
\dontrun{GetModuleList()}
}

