% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetModules}
\alias{GetModules}
\title{A function to apply GetModule to a list correctly.}
\usage{
GetModules(modules)
}
\arguments{
\item{modules}{A list from CheckModList() given details for
one or more modules.}
}
\description{
A function to apply GetModule to a list correctly.
}

