% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPackage.R
\name{GetPackage}
\alias{GetPackage}
\title{Helper to install (if needed) and load a package}
\usage{
GetPackage(package)
}
\arguments{
\item{package}{A character vector of packages to load}
}
\description{
Given a package name, either as a string or object,
load the package if it exists, else install it from
CRAN and then load
}
\examples{
\dontrun{

GetPackage('gam')

}
}
