% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{LoadModule}
\alias{LoadModule}
\title{A function to load a module function from url or disk.}
\usage{
LoadModule(module)
}
\arguments{
\item{module}{A string that describes the location of the R file. Can be a
a full URL or a path to a local file.}
}
\value{
Name of the function. Adds function to global namespace.
}
\description{
Loads a module function into the global environment ready to be used in a
 zoon workflow. This function is mostly for use while developing modules.
 Workflows run with modules defined locally are no longer reproducible and
 so are discouraged and will be tagged as 'unreproducible'.
}
