% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modelLR}
\alias{modelLR}
\title{model module to fit a simple logistic regression model}
\usage{
modelLR(df)
}
\arguments{
\item{df}{A dataframe, the output from a process module}
}
\value{
A model object with a valid predict method
}
\description{
model module to fit a simple logistic regression model
}

