% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modelRF}
\alias{modelRF}
\title{model module to fit a simple RandomForest classification model}
\usage{
modelRF(df)
}
\arguments{
\item{df}{A dataframe, the output from a process module}
}
\value{
A model object with a valid predict method
}
\description{
model module to fit a simple RandomForest classification model
}

