% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleArguments.R
\name{ModuleArguments}
\alias{ModuleArguments}
\title{ModuleArguments}
\usage{
ModuleArguments(module)
}
\arguments{
\item{module}{string giving the name of the module}
}
\value{
A list of arguments to that module
}
\description{
Produce list of module arguments
}
