% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ModuleHelp}
\alias{ModuleHelp}
\title{ModuleHelp}
\usage{
ModuleHelp(module)
}
\arguments{
\item{module}{The name of a zoon module (as a string)}
}
\value{
Prints the help page to screen.
}
\description{
Returns the help file for a zoon module
Currently only text help is implemented.
}
\seealso{
\code{\link{GetModuleList}}
}

