% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{occurrence}
\alias{covariate}
\alias{model}
\alias{occurrence}
\alias{output}
\alias{process}
\title{Accessor functions for getting module outputs from a workflow object}
\usage{
occurrence(workflow)

covariate(workflow)

process(workflow)

model(workflow)

output(workflow)
}
\arguments{
\item{workflow}{A workflow object}
}
\description{
These functions access the output from each module type. If workflows
  are split using list, they will return a list with the output of 
  each seperate workflow being one element of the list.
}
\examples{
\dontrun{
work1 <- workflow(occurrence = UKAnophelesPlumbeus,
                 covariate  = UKAir,
                 process    = OneHundredBackground,
                 model      = list(LogisticRegression, LogisticRegression),
                 output     = PrintMap)

occurrence(work1)
covariate(work1)
process(work1)
model(work1)
model(work1)[[1]]
output(work1)
}
}

