% Generated by roxygen2 (4.0.1): do not edit by hand
\name{occurrenceCp}
\alias{occurrenceCp}
\title{occurrence module to grab *Culex rajah* (a mosquito) occurrence (i.e.
      presence-only) data from GBIF, in the area bounded by extent.
      Perhaps this should have temporal interval too for future-proofing?}
\usage{
occurrenceCp(extent)
}
\arguments{
\item{extent}{A numeric vetor of length 4 giving the coordinates of the
      rectangular region within which to carry out the analysis, in the
      order: xmin, xmax, ymin, ymax.}
}
\value{
a dataframe with four columns:
      value - a numeric value which may give 1 for presences, 0 for absences
      or a positive integer for count data
      type - a character value saying what is in the value column
      lon - the longitude of the record
      lat - the latitutude of the record
}
\description{
occurrence module to grab *Culex rajah* (a mosquito) occurrence (i.e.
      presence-only) data from GBIF, in the area bounded by extent.
      Perhaps this should have temporal interval too for future-proofing?
}

