% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PasteAndDep}
\alias{PasteAndDep}
\title{PasteAndDep}
\usage{
PasteAndDep(x)
}
\description{
Paste and deparse. Helper to format substituted args. If arguments were
  chained or listed then substitute gives a list. We want to paste it back together.
}

