% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.zoonWorkflow.R
\name{plot.zoonWorkflow}
\alias{plot.zoonWorkflow}
\title{Plot a schematic illustration of a zoon workflow structure.}
\usage{
\method{plot}{zoonWorkflow}(x, ...)
}
\arguments{
\item{x}{an object of class zoonWorkflow}

\item{\dots}{currently ignored}
}
\description{
Opens a graphics device and produces a plot of the workflow
structure and modules used.
}
