% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.zoonSummary.R
\name{print.Summary}
\alias{print.Summary}
\alias{print.zoonSummary}
\title{A function to print zoon workflow summaries}
\usage{
\method{print}{zoonSummary}(x, ...)
}
\arguments{
\item{x}{object of class zoonSummary}

\item{\dots}{currently ignored}
}
\description{
Prints a simple zoonSummary object to console
}

