% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.zoonCitation.R
\name{print.zoonCitation}
\alias{print.zoonCitation}
\title{A function to print zoonCitation}
\usage{
\method{print}{zoonCitation}(x, ...)
}
\arguments{
\item{x}{object of class zoonCitation}

\item{\dots}{currently ignored}
}
\description{
Prints a zoonCitation object to console
 giving easy access to module citations
}
