% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.zoonWorkflow.R
\name{print.zoonWorkflow}
\alias{print.zoonWorkflow}
\title{A function to print a zoonWorkflow object}
\usage{
\method{print}{zoonWorkflow}(x, ...)
}
\arguments{
\item{x}{object of class zoonWorkflow}

\item{\dots}{currently ignored}
}
\description{
The function returns a very simple output detailing the function call.
}
