% Generated by roxygen2 (4.0.1): do not edit by hand
\name{processA}
\alias{processA}
\title{process module to generate up to 100 background records at random in
     cells of ras and return these along with the ppresence only data.}
\usage{
processA(occ, ras)
}
\arguments{
\item{occ}{Occurrence data, the output from an occurrence module}

\item{ras}{Covariate data, the output from a covariate module}
}
\value{
Dataframe with at least 5 columns
      value - a numeric value which may give 1 for presences, 0 for absences
      or a positive integer for count data
      type - a character value saying what is in the value column
      lon - the longitude of the record
      lat - the latitutude of the record
      columns 5-n - the values of the covariates for each records (the names of
              these columns should correspond exactly to the names of the
              layers in ras)
}
\description{
process module to generate up to 100 background records at random in
     cells of ras and return these along with the ppresence only data.
}

