% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RerunWorkflow}
\alias{RerunWorkflow}
\title{Rerun a workflow object.}
\usage{
RerunWorkflow(workflow, from = NULL)

RerunWorkflow(workflow, from = NULL)
}
\arguments{
\item{workflow}{A zoonWorkflow object from a previous zoon analysis}

\item{from}{Which modules should be run. Takes an integer and then
 starts from that module with the order being the argument order
 (i.e. process is the third module). Typically this might be used to run
 an analysis using the data in the workflow with from=3 if the data
 collection modules are for reading in local data.}

\item{workflow}{A zoonWorkflow object from a previous zoon analysis}

\item{from}{Which modules should be run. If NULL (default), run from the
 first NULL output (i.e. where the workflow broke). Otherwise takes an
 integer and runs from that module.}
}
\value{
A list with the results of each module and a copy of the
 call used to execute the workflow (

A list with the results of each module and a copy of the
 call used to execute the workflow.
}
\description{
Takes a workflow object and reruns it.

Takes a workflow object and reruns it.
}
\examples{
\dontrun{
w <- workflow(UKAnophelesPlumbeus, UKAir,
              OneHundredBackground,
              LogisticRegression,
              SameTimePlaceMap)

RerunWorkflow(w)
}
\dontrun{
w <- workflow(UKAnophelesPlumbeus, UKAir,
              OneHundredBackground,
              LogisticRegression,
              SameTimePlaceMap)

RerunWorkflow(w)
}
}

