% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{RunModels}
\alias{RunModels}
\title{RunModels}
\usage{
RunModels(df, modelFunction, paras, workEnv)
}
\arguments{
\item{df}{Dataframe from process module. Should contain columns value, type,
lon, lat and fold, a number indicating which cross validation set a
datapoint is in. If fold is 0 then this is considered external validation
data If all data is 0 or 1, then no cross validation is run.}

\item{modelFunction}{String giving the name of the model function which is in
turn the name of the module.}

\item{paras}{All other parameters that should be passed to the model
function. i.e. model[[1]]$paras}

\item{workEnv}{The environment name of the workflow call environment.}
}
\value{
A list of length 2 containing the model trained on all data and a
  data.frame which contains value, type, fold, lon, lat, predictions and then
  all environmental variables.
}
\description{
A function to train and predict crossvalidation folds and train one full
model and predict any external validation data. This function is primarily
used internally but can be used when running workflows interactively (see
vignette \code{basic zoon useage})
}
