% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SortArgs}
\alias{SortArgs}
\title{SortArgs}
\usage{
SortArgs(occSub, covSub, proSub, modSub, outSub, forceReproducible)
}
\description{
Helper to take substituted args from workflow call and paste them into
  a runeable workflow function.
}

