% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.zoonWorkflow.R
\name{summary.zoonWorkflow}
\alias{summary.zoonWorkflow}
\title{A function to summarize the output of a zoon workflow}
\usage{
\method{summary}{zoonWorkflow}(workflow)
}
\arguments{
\item{workflow}{A zoon workflow object}
}
\value{
A list that summarises the workflow
}
\description{
The function currently just outputs the dimensions and a few other
  parts of a workflow. This could certainly be developed to include useful
  bit of information from each module type e.g. the extent of rasters, the
  type of model object etc.
}

