% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformCRS.R
\name{TransformCRS}
\alias{TransformCRS}
\title{Change the CRS of occurrence or covariate data}
\usage{
TransformCRS(data, module)
}
\arguments{
\item{data}{The object returned by an occurrence or covariate module.}

\item{module}{The module the object came from, one of 'occurrence' or 'covariate'}
}
\value{
The same object as in data, but with CRS changed
}
\description{
Takes a dataframe returned by an occurrence module or a raster object from a covariate module and converts the CRS to lat/long so tht everything works together.
}
\examples{
\dontrun{
w <- workflow(UKAnophelesPlumbeus,
              UKAir,
              Background(n = 70), 
              LogisticRegression,
              PrintMap)

w2 <- ChangeWorkflow(w,
                     output = PrintMap)
}
}

