% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{zoon}
\alias{zoon}
\alias{zoon-package}
\title{Zoon: A package for comparing multple SDM models, good model diagnostics
     and better reproducibility}
\description{
Zoon: A package for comparing multple SDM models, good model diagnostics
     and better reproducibility
}

