% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZoonCitation.R
\name{ZoonCitation}
\alias{ZoonCitation}
\title{ZoonCitation}
\usage{
ZoonCitation(ModuleName)
}
\arguments{
\item{ModuleName}{string giving the name of the module}
}
\value{
A \code{zoonCitation} object - A list of elements
that make up the citation
}
\description{
How to cite Zoon Modules in publications
}
\seealso{
\code{\link{print.zoonCitation}}
}
