% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZoonModuleParse.R
\name{ZoonModuleParse}
\alias{ZoonModuleParse}
\title{Parse a module file to read roxygen tags}
\usage{
ZoonModuleParse(modulePath)
}
\arguments{
\item{modulePath}{The path to a zoon module .R script}
}
\value{
Named list of roxygen tags
}
\description{
This is a stand in for the parse_file function in roxygen which
is not exported.
}
