% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/BuildModule.R
\name{BuildModule}
\alias{BuildModule}
\title{BuildModule}
\usage{
BuildModule(object, type, dir = ".", title = "", description = "",
  details = "", author = "", email = "", paras = NULL)
}
\arguments{
\item{object}{A function that will be made into a module file.}

\item{type}{A string that defines the type of module. Possible module types
are occurrence, covariate, process, model, diagnostic and output.}

\item{dir}{The directory to put the module into (defaults to the
working directory.}

\item{title}{A short description of the module.}

\item{description}{(required) A single string giving a full description of the module.}

\item{details}{(optional) A single string giving details of the module.}

\item{author}{(required) String giving the author(s) name(s)}

\item{email}{(required) String giving the correspondance address for the module.}

\item{paras}{A list of the form
list(parameterName = 'Parameter description.',
anotherParameter = 'Another descriptions.')
This is required if the module takes non-default arguements}
}
\value{
Name of the module. Outputs a file
}
\description{
Turn a function in the namespace into a module.
Will later add functions to upload module to figshare etc.
And add testing that the module name is unique.
}

