% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BuildModule.R
\name{BuildModule}
\alias{BuildModule}
\title{BuildModule}
\usage{
BuildModule(object, type, dir = ".", title = "", description = "",
  author = "", email = "", paras = NULL)
}
\arguments{
\item{object}{A function that will be made into a module file.}

\item{type}{A string that defines the type of module. Possible module types
are occurrence, covariate, process, model, diagnostic and output.}

\item{dir}{The directory to put the module into (defaults to the
working directory.}

\item{title}{A short description of the module.}

\item{description}{A single string giving a full description of the model.}

\item{author}{String giving the author(s) name(s)}

\item{email}{Stirng giving the correspondance address for the module.}

\item{paras}{A list of the form
list(parameterName = 'Parameter description.',
  anotherParameter = 'Another descriptions.')}
}
\value{
NULL. Outputs a file
}
\description{
Turn a function in the namespace into a module.
Will later add functions to upload module to figshare etc.
And add testing that the module name is unique.
}

