% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{Chain}
\alias{Chain}
\title{Chain modules together}
\usage{
Chain(...)
}
\arguments{
\item{...}{List of modules to be chained.}
}
\description{
\code{Chain} combines multiple modules of the same module type such that they
are executed sequentially and their outputs combined. For example, process
modules may be \code{Chain}ed to carry out successive processing operations.
By contrast, \code{list}ing modules of the same type would split the workflow
into multiple parallel workflows, each using a different module at this step.
}
\details{
Similarly for occurrence or covariate modules the datasets are joined (row-
or layer-wise) whereas \code{list} would carry out separate analyses. Model
and output modules may not be chained. Developers should note that this
function is not actually used - calls using \code{Chain} are parsed by
workflow, with behaviour similar to this function.
}
