% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{CheckModList}
\alias{CheckModList}
\title{CheckModList}
\usage{
CheckModList(x)
}
\arguments{
\item{x}{A 'call' or 'character' from substitute(occurrence) etc. So either
quoted which yields a character substitute('module1') or unquotes which
yields a call substitute(module1).}
}
\description{
Helper to sort module arguments into lists with common structure.
  Want any input to end up as
  list(module = moduleName, paras = list(paraName1 = para, paraName2 = 2)).
  See tests/testthat/testZoon.R for list of potential input.
}

