% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{ExtractAndCombData}
\alias{ExtractAndCombData}
\title{ExtractAndCombData}
\usage{
ExtractAndCombData(occurrence, ras)
}
\arguments{
\item{occurrence}{A data frame from an occurrence module}

\item{ras}{Environmental raster layer, stack or brick.}
}
\description{
Simply extract covariates from rasters and combine with occurrence data.
This function is primarily used internally but can be used when
running workflows interactively (see vignette \code{basic zoon useage})
}
