% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FormatModuleList}
\alias{FormatModuleList}
\title{FormatModuleList}
\usage{
FormatModuleList(x)
}
\description{
Little helper to format module lists. Want to return a list
  newList$module is the module name and newList$paras is a list
  of the parameters given for the module.
}

