% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{FormatModuleList}
\alias{FormatModuleList}
\title{FormatModuleList}
\usage{
FormatModuleList(x)
}
\arguments{
\item{x}{An object of class 'name' or 'call' e.g. module1 or module1(k=2).}
}
\description{
Little helper to format module lists. Want to return a list
  newList$module is the module name and newList$paras is a list
  of the parameters given for the module.
}

