% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetModule}
\alias{GetModule}
\title{A function to get a module function.
It assumes the module is in github.com/zoonproject unless it is a full url
Assigns the function to the global environment.}
\usage{
GetModule(module)
}
\arguments{
\item{module}{A string that describes the location of the R file. Can be a
     module name assuming the module is in github.com/zoonproject/modules.
     Otherwise can be a full URL or a local file.}
}
\value{
Name of the function. Adds function to global namespace.
}
\description{
A function to get a module function.
It assumes the module is in github.com/zoonproject unless it is a full url
Assigns the function to the global environment.
}

