% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{GetModule}
\alias{GetModule}
\title{A function to get a module function.}
\usage{
GetModule(module, forceReproducible)
}
\arguments{
\item{module}{A string that describes the location of the R file. Can be a
module name assuming the module is in global namespace or
github.com/zoonproject/modules. Otherwise can be a full URL or a local
file.}

\item{forceReproducible}{Do we want to force the function to get modules from
     the zoon repo, even if they exist in the global namespace.}
}
\value{
Name of the function. Function is run with workflow and new module
 function is added to workflow environment.
}
\description{
Checks for the module in the global namespace, then the zoon repo. Then reads
 the module source and loads the function.
}

