% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetModuleList}
\alias{GetModuleList}
\title{GetModuleList}
\usage{
GetModuleList(moduleType = "all")
}
\arguments{
\item{moduleType}{The type of modules (character vector) to be listed out of
occurrence, covariate, process, model and output.}
}
\value{
A list with all module names.
}
\description{
Get a list of all the modules available on the github repo.
}
\examples{
\dontrun{GetModuleList()}
}

