% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetModuleList}
\alias{GetModuleList}
\title{GetModuleList}
\usage{
GetModuleList(renew = FALSE)
}
\arguments{
\item{renew}{Download from github even if we already have a module list.}
}
\value{
A list with all module names.
}
\description{
Get a list of all the modules available on the github repo.
}
\examples{
\dontrun{GetModuleList()}
}

