% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetModules}
\alias{GetModules}
\title{A function to apply GetModule to a list correctly.}
\usage{
GetModules(modules, forceReproducible)
}
\arguments{
\item{modules}{A list from CheckModList() given details for
one or more modules.}

\item{forceReproducible}{Logical to determine whether the modules should be
taken from repo even if they exist locally to enforce reproducibility.}
}
\description{
A function to apply GetModule to a list correctly.
}

