% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFunctions.R
\name{ModuleHelp}
\alias{ModuleHelp}
\title{ModuleHelp}
\usage{
ModuleHelp(module)
}
\arguments{
\item{module}{The name of a zoon module}
}
\value{
Prints the help page to screen.
}
\description{
Returns the help file for a zoon module.
}
\seealso{
\code{\link{GetModuleList}}
}
