% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ModuleOptions}
\alias{GetModuleList}
\alias{ModuleOptions}
\title{Turns named options into list ready to be used by workflow()}
\usage{
GetModuleList(type = "NULL")

ModuleOptions(module, ...)
}
\arguments{
\item{module}{The module name or URL.}

\item{...}{Any other parameters or options needed by that  module.
          All extra options must be named i.e. ModuleOptions('x', x=1)
          Not ModuleOptions('x', 1).}

\item{module}{The module name or URL.}

\item{...}{Any other parameters or options needed by that  module.
          All extra options must be named i.e. ModuleOptions('x', x=1)
          Not ModuleOptions('x', 1).}
}
\value{
A list with all module options and the module name/URL in.

A list with all module options and the module name/URL in.
}
\description{
Turns named options into list ready to be used by workflow()

Turns named options into list ready to be used by workflow()
}
\examples{
print('No examples yet')
print('No examples yet')
}

