% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{SplitArgs}
\alias{SplitArgs}
\title{Split a string into a module and it's arguments}
\usage{
SplitArgs(string)
}
\arguments{
\item{string}{A string of the form "moduleName" or
 "moduleName(parameter = 2, parameter2 = 3)"}
}
\description{
A function that takes a string (from workflow$call) and splits it into a
  module name and it's arguments.
}

