% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoonHelpers.R
\name{SplitCall}
\alias{SplitCall}
\title{SplitCall}
\usage{
SplitCall(call)
}
\arguments{
\item{call}{A character string of a valid zoon workflow call.}
}
\description{
Helper to split a character string workflow call, as inherited from zoonWorkflow
  into it's constituent arguments
}

