% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.zoonWorkflow.R
\name{summary.zoonWorkflow}
\alias{summary.zoonWorkflow}
\title{A function to summarize the output of a zoon workflow}
\usage{
\method{summary}{zoonWorkflow}(object, ...)
}
\arguments{
\item{object}{A zoonWorkflow object}

\item{\dots}{currently ignored}
}
\value{
A list of length 2, first the data summaries and then the
model summaries. each of these is as long as the coressponding
elements in the zoonWorkflow object
}
\description{
The function currently outputs the result of summary() on each
occurence element and each model element
}

