% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.zoonWorkflow.R
\name{summary.zoonWorkflow}
\alias{summary.zoonWorkflow}
\title{A function to summarize the output of a zoon workflow}
\usage{
\method{summary}{zoonWorkflow}(object, ...)
}
\arguments{
\item{object}{A zoonWorkflow object}

\item{\dots}{currently ignored}
}
\value{
Path to HTML file. Associated images will be in the same directory.
}
\description{
Renders and opens an HTML report
}
