% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformCRS.R
\name{TransformCRS}
\alias{TransformCRS}
\title{Change the CRS of occurrence data}
\usage{
TransformCRS(occurrence, ras_projection)
}
\arguments{
\item{occurrence}{The output of an occurrence module}

\item{ras_projection}{The projection of a covariate layer as a character
(from projection())}
}
\value{
The same object as in occurrence, but with CRS changed as needed
}
\description{
Takes a dataframe returned by an occurrence module or a raster object from a
covariate module and converts the CRS to lat/long so that everything works
together.
}
