% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{UKAirRas}
\alias{UKAirRas}
\title{UK Air temperature raster layer.}
\format{Raster layer}
\source{
NCEP
}
\description{
This is an example environmental covariate raster layer. It is
  surface temperatures for the UK taken from NCEP
}
\author{
Tim Lucas September 2014
}
