% Generated by roxygen2 (4.0.2): do not edit by hand
\name{workflow}
\alias{workflow}
\title{Run a full workflow.}
\usage{
workflow(occurrence, covariate, process, model, output)
}
\arguments{
\item{occurrence}{The name of the module to be used to get occurence data}

\item{covariate}{The name of the module to be used to get covariate
data.}

\item{process}{The name of the module to be used to process the data}

\item{model}{The name of the SDM model module to be used}

\item{output}{The name of the module to be used to map output}
}
\value{
A list with the results of each module and a copy of the
 code used to execute the workflow (what's there now should be source-able
 though I'm sure there is a much neater approach than the one I took - the
 ultimate aim would be a much nicer way of enhancing reproducibility).
}
\description{
This is the main function of zoon. The arguments should specify at least five
   modules, at least one of each type.
}
\examples{
# run a workflow, using the logistic regression model
work1 <- workflow(occurMod = 'UKAnophelesPlumbeus',
                covarMod = 'UKAir',
                procMod = 'OneHundredBackground',
                modelMod = 'LogisticRegression',
                outMod = 'SameTimePlaceMap')

str(work1, 1)

work2 <- workflow('UKAnophelesPlumbeus', 'UKAir', 'OneHundredBackground',
          'RandomForest', 'PrintMap')
}

