% Generated by roxygen2 (4.0.1): do not edit by hand
\name{workflow}
\alias{workflow}
\title{workflow wrapper function}
\usage{
workflow(ext, occurrenceFn, covariateFn, processFn, modelFn, mapFn)
}
\arguments{
\item{ext}{A numeric vector of length 4 giving the coordinates of the
rectangular region within which to carry out the analysis, in the
order: xmin, xmax, ymin, ymax.}

\item{occurrenceFn}{The name of the function (module) to be used to get occurence data}

\item{covariateFn}{The name of the function (module) to be used to get covariate data}

\item{processFn}{The name of the function (module) to be used to process the data}

\item{modelFn}{The name of the SDM model function (module) to be used}

\item{mapFn}{The name of the function (module) to be used to map output}
}
\value{
A list with the extent, the results of each module and a copy of the
      code used to execute the workflow (what's there now should be source-able
      though I'm sure there is a much neater approach than the one I took - the
      ultimate aim would be a much nicer way of enhancing reproducibility).
}
\description{
workflow wrapper function
}
\examples{
x <- 2+2
}

