% Generated by roxygen2 (4.0.1): do not edit by hand
\name{workflow}
\alias{workflow}
\title{workhorse}
\usage{
workflow(occurMod, covarMod, procMod, modelMod, outMod)
}
\arguments{
\item{occurMod}{The name of the function (module) to be used to get occurence data}

\item{covariate.module}{The name of the function (module) to be used to get covariate data}

\item{process.module}{The name of the function (module) to be used to process the data}

\item{model.module}{The name of the SDM model function (module) to be used}

\item{output.module}{The name of the function (module) to be used to map output}
}
\value{
A list with the results of each module and a copy of the
      code used to execute the workflow (what's there now should be source-able
      though I'm sure there is a much neater approach than the one I took - the
      ultimate aim would be a much nicer way of enhancing reproducibility).
}
\description{
workhorse
}
\examples{
# run a workflow, using the logistic regression model
\dontrun{ans1 <- workflow(occurMod = 'AnophelesPlumbeus',
                covarMod = 'AirNCEP',
                procMod = 'OneHundredBackground',
                modelMod = 'LogisticRegression',
                outMod = 'SameTimePlaceMap')

str(ans1, 1)


# plot the resulting maps

plot(ans1$map,
    zlim = c(0, 1),
    main = 'LR')


points(ans1$occ[, 1:2],
      pch = 16,
      cex = 0.3)
}
}

