% Generated by roxygen2 (4.0.1): do not edit by hand
\name{workflow}
\alias{workflow}
\title{workflow wrapper function}
\usage{
workflow(extent, occurrence.module, covariate.module, process.module,
  model.module, map.module)
}
\arguments{
\item{extent}{A numeric vector of length 4 giving the coordinates of the
rectangular region within which to carry out the analysis, in the
order: xmin, xmax, ymin, ymax.}

\item{occurrence.module}{The name of the function (module) to be used to get occurence data}

\item{covariate.module}{The name of the function (module) to be used to get covariate data}

\item{process.module}{The name of the function (module) to be used to process the data}

\item{model.module}{The name of the SDM model function (module) to be used}

\item{map.module}{The name of the function (module) to be used to map output}
}
\value{
A list with the extent, the results of each module and a copy of the
      code used to execute the workflow (what's there now should be source-able
      though I'm sure there is a much neater approach than the one I took - the
      ultimate aim would be a much nicer way of enhancing reproducibility).
}
\description{
workflow wrapper function
}
\examples{
# define the extent in lat and long
uk.extent <- c(xmin = -10,
             xmax = 10,
             ymin = 45,
             ymax = 65)

# run a workflow, using the logistic regression model
\dontrun{ans1 <- workflow(extent = uk.extent,
                occurrence.module = 'AnophelesPlumbeus',
                covariate.module = 'AirNCEP',
                process.module = 'OneHundredBackground',
                model.module = 'LogisticRegression',
                map.module = 'SameTimePlaceMap')

str(ans1, 1)


# plot the resulting maps

plot(ans1$map,
    zlim = c(0, 1),
    main = 'LR')


points(ans1$occ[, 1:2],
      pch = 16,
      cex = 0.3)
}
}

