% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoon.R
\docType{package}
\name{zoon}
\alias{zoon}
\alias{zoon-package}
\title{Zoon: A package for comparing multple SDM models, good model diagnostics
     and better reproducibility}
\description{
Zoon: A package for comparing multple SDM models, good model diagnostics
     and better reproducibility
}

